matmathutils and markerutils merge wip
